ai_ecom_assistant/requirements.txt
# Backend
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0

# AI Models
transformers==4.35.2
torch==2.1.0
accelerate==0.24.1
bitsandbytes==0.41.1
diffusers==0.24.0
pillow==10.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Payments
stripe==5.5.0
